requirements.txt
streamlit
pandas
numpy
xgboost
scikit-learn
pyupbit
pybithumb
python-dotenv
ta
matplotlib
plotly
watchdog
